ui/requirements.txt
flask
WTForms
Flask-WTF
wtforms_components
requests
python-dotenv